deprecate meteodatalab but keep in destaggering
