+++ src/anemoi_plugins_meteoswiss/transform/filters/__init__.py
@@ -1,20 +1,16 @@
-from .clipping import ClipLateralBoundaries
 from .destaggering import Destagger
 from .geopotential_from_height import GeopotentialFromHeight
 from .grid import AssignGrid
-from .omega_from_w import OmegaFromW
 from .time_processing import AverageFluxToCumulativeQuantity
 from .vertical_interpolation import ModelToPressureLevel
 
 __all__ = [
     "AverageFluxToCumulativeQuantity",
-    "ClipLateralBoundaries",
     "Destagger",
     "AssignGrid",
     "ModelToPressureLevel",
     "Interp2Grid",
     "InterpNAFilter",
     "Interp2Res",
-    "OmegaFromW",
     "GeopotentialFromHeight",
 ]
